Simply test structure
