+++ tests/Feature/LoginTest.php
@@ -9,7 +9,7 @@
  * file that was distributed with this source code.
  */
 
-namespace Tests\Web;
+namespace Tests\Feature;
 
 use Auth;
 use Tests\Foundation\TestCase;
